removed dead code, split large files into smaller files, reduced repeated code.
