+++ app/src/main/java/com/example/stitchcounterv3/feature/doublecounter/DoubleCounterScreen.kt
@@ -1,17 +1,7 @@
 package com.example.stitchcounterv3.feature.doublecounter
 
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Info
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -29,6 +19,8 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.example.stitchcounterv3.feature.navigation.RootNavGraph
 import com.example.stitchcounterv3.feature.sharedComposables.AdaptiveLayout
+import com.example.stitchcounterv3.feature.sharedComposables.ProjectDetailsFAB
+import com.example.stitchcounterv3.feature.sharedComposables.ResetConfirmationDialog
 import com.ramcosta.composedestinations.annotation.Destination
 
 @RootNavGraph
@@ -82,21 +74,9 @@ fun DoubleCounterScreen(
                     actions = actions,
                     topBarContent = if (state.id > 0 && onNavigateToDetail != null) {
                         {
-                            FloatingActionButton(
-                                onClick = {
-                                    onNavigateToDetail(state.id)
-                                },
-                                modifier = Modifier
-                                    .padding(start = 16.dp)
-                                    .size(40.dp),
-                                containerColor = MaterialTheme.colorScheme.primaryContainer
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.Info,
-                                    contentDescription = "Project details",
-                                    modifier = Modifier.size(20.dp)
-                                )
-                            }
+                            ProjectDetailsFAB(
+                                onClick = { onNavigateToDetail(state.id) }
+                            )
                         }
                     } else null
                 )
@@ -107,21 +87,9 @@ fun DoubleCounterScreen(
                     actions = actions,
                     topBarContent = if (state.id > 0 && onNavigateToDetail != null) {
                         {
-                            FloatingActionButton(
-                                onClick = {
-                                    onNavigateToDetail(state.id)
-                                },
-                                modifier = Modifier
-                                    .padding(start = 16.dp)
-                                    .size(40.dp),
-                                containerColor = MaterialTheme.colorScheme.primaryContainer
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.Info,
-                                    contentDescription = "Project details",
-                                    modifier = Modifier.size(20.dp)
-                                )
-                            }
+                            ProjectDetailsFAB(
+                                onClick = { onNavigateToDetail(state.id) }
+                            )
                         }
                     } else null
                 )
@@ -134,52 +102,26 @@ fun DoubleCounterScreen(
             CounterType.STITCH -> "Stitches"
             CounterType.ROW -> "Rows/Rounds"
         }
-        AlertDialog(
-            onDismissRequest = { resetDialogType = null },
-            title = { Text("Reset $counterName Counter?") },
-            text = { Text("Are you sure you want to reset the $counterName counter to 0?") },
-            confirmButton = {
-                TextButton(
-                    onClick = {
-                        viewModel.reset(type)
-                        resetDialogType = null
-                    }
-                ) {
-                    Text("Reset")
-                }
+        ResetConfirmationDialog(
+            title = "Reset $counterName Counter?",
+            message = "Are you sure you want to reset the $counterName counter to 0?",
+            onConfirm = {
+                viewModel.reset(type)
+                resetDialogType = null
             },
-            dismissButton = {
-                TextButton(
-                    onClick = { resetDialogType = null }
-                ) {
-                    Text("Cancel")
-                }
-            }
+            onDismiss = { resetDialogType = null }
         )
     }
 
     if (showResetAllDialog) {
-        AlertDialog(
-            onDismissRequest = { showResetAllDialog = false },
-            title = { Text("Reset All Counters?") },
-            text = { Text("Are you sure you want to reset both Stitches and Rows/Rounds counters to 0?") },
-            confirmButton = {
-                TextButton(
-                    onClick = {
-                        viewModel.resetAll()
-                        showResetAllDialog = false
-                    }
-                ) {
-                    Text("Reset All")
-                }
+        ResetConfirmationDialog(
+            title = "Reset All Counters?",
+            message = "Are you sure you want to reset both Stitches and Rows/Rounds counters to 0?",
+            onConfirm = {
+                viewModel.resetAll()
+                showResetAllDialog = false
             },
-            dismissButton = {
-                TextButton(
-                    onClick = { showResetAllDialog = false }
-                ) {
-                    Text("Cancel")
-                }
-            }
+            onDismiss = { showResetAllDialog = false }
         )
     }
 }
