+++ app/src/main/java/com/example/stitchcounterv3/feature/doublecounter/DoubleCounterViewModel.kt
@@ -9,7 +9,6 @@ import com.example.stitchcounterv3.domain.model.Project
 import com.example.stitchcounterv3.domain.model.ProjectType
 import com.example.stitchcounterv3.domain.usecase.GetProject
 import com.example.stitchcounterv3.domain.usecase.UpsertProject
-import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.Job
@@ -28,7 +27,6 @@ data class DoubleCounterUiState(
     val stitchCounterState: CounterState = CounterState(),
     val rowCounterState: CounterState = CounterState(),
     val totalRows: Int = 0,
-    val isLoading: Boolean = false,
 ) {
     val rowProgress: Float? = if (totalRows > 0) {
         (rowCounterState.count.toFloat() / totalRows.toFloat()).coerceIn(0f, 1f)
@@ -46,7 +44,7 @@ enum class CounterType {
 open class DoubleCounterViewModel @Inject constructor(
     private val getProject: GetProject,
     private val upsertProject: UpsertProject,
-) : ViewModel(), DoubleCounterActions {
+) : ViewModel() {
 
     private val _uiState = MutableStateFlow(DoubleCounterUiState())
     open val uiState: StateFlow<DoubleCounterUiState> = _uiState.asStateFlow()
@@ -63,7 +61,9 @@ open class DoubleCounterViewModel @Inject constructor(
                 resetState()
                 return@launch
             }
-            _uiState.update { currentState -> currentState.copy(isLoading = true) }
+            if (_uiState.value.id == projectId) {
+                return@launch
+            }
             val project = getProject(projectId)
             if (project != null) {
                 _uiState.update { currentState ->
@@ -78,12 +78,9 @@ open class DoubleCounterViewModel @Inject constructor(
                             count = project.rowCounterNumber,
                             adjustment = AdjustmentAmount.entries.find { it.adjustmentAmount == project.rowAdjustment } ?: AdjustmentAmount.ONE
                         ),
-                        totalRows = project.totalRows,
-                        isLoading = false
+                        totalRows = project.totalRows
                     )
                 }
-            } else {
-                _uiState.update { currentState -> currentState.copy(isLoading = false) }
             }
         }
     }
@@ -103,10 +100,10 @@ open class DoubleCounterViewModel @Inject constructor(
         triggerAutoSave()
     }
 
-    override fun increment(type: CounterType) = updateCounter(type) { it.increment() }
-    override fun decrement(type: CounterType) = updateCounter(type) { it.decrement() }
-    override fun reset(type: CounterType) = updateCounter(type) { it.reset() }
-    override fun changeAdjustment(type: CounterType, value: AdjustmentAmount) = 
+    fun increment(type: CounterType) = updateCounter(type) { it.increment() }
+    fun decrement(type: CounterType) = updateCounter(type) { it.decrement() }
+    fun reset(type: CounterType) = updateCounter(type) { it.reset() }
+    fun changeAdjustment(type: CounterType, value: AdjustmentAmount) = 
         updateCounter(type) { it.copy(adjustment = value) }
     
     private fun triggerAutoSave() {
@@ -145,31 +142,8 @@ open class DoubleCounterViewModel @Inject constructor(
     fun attemptDismissal() {
         viewModelScope.launch {
             autoSaveJob?.cancel()
-            val s = _uiState.value
-            val existingProject = if (s.id > 0) getProject(s.id) else null
-            val project = Project(
-                id = s.id,
-                type = ProjectType.DOUBLE,
-                title = existingProject?.title ?: "",
-                stitchCounterNumber = s.stitchCounterState.count,
-                stitchAdjustment = s.stitchCounterState.adjustment.adjustmentAmount,
-                rowCounterNumber = s.rowCounterState.count,
-                rowAdjustment = s.rowCounterState.adjustment.adjustmentAmount,
-                totalRows = existingProject?.totalRows ?: s.totalRows,
-                imagePath = existingProject?.imagePath
-            )
-            val newId = upsertProject(project).toInt()
-            if (s.id == 0 && newId > 0) {
-                _uiState.update { currentState -> currentState.copy(id = newId) }
-            }
-            _dismissalResult.send(DismissalResult.Allowed)
-        }
-    }
-
-    fun saveAndGoBack(navigator: DestinationsNavigator) {
-        viewModelScope.launch {
             save()
-            navigator.popBackStack()
+            _dismissalResult.send(DismissalResult.Allowed)
         }
     }
 
@@ -177,7 +151,7 @@ open class DoubleCounterViewModel @Inject constructor(
         _uiState.update { _ -> DoubleCounterUiState() }
     }
 
-    override fun resetAll() {
+    fun resetAll() {
         reset(CounterType.STITCH)
         reset(CounterType.ROW)
     }
