+++ app/src/main/java/com/example/stitchcounterv3/feature/library/LibraryTopBars.kt
@@ -1,0 +1,99 @@
+package com.example.stitchcounterv3.feature.library
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.SelectAll
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LibraryTopBar(
+    onEnterMultiSelect: () -> Unit,
+    hasProjects: Boolean = true
+) {
+    TopAppBar(
+        title = {
+            Text("Library")
+        },
+        actions = {
+            if (hasProjects) {
+                IconButton(onClick = onEnterMultiSelect) {
+                    Icon(
+                        imageVector = Icons.Default.SelectAll,
+                        contentDescription = "Select multiple"
+                    )
+                }
+            }
+        }
+    )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MultiSelectTopBar(
+    selectedCount: Int,
+    totalCount: Int,
+    onSelectAll: () -> Unit,
+    onClearSelection: () -> Unit,
+    onDelete: () -> Unit,
+    onCancel: () -> Unit
+) {
+    TopAppBar(
+        title = {
+            Text(
+                text = if (selectedCount > 0) {
+                    "$selectedCount selected"
+                } else {
+                    "Select projects"
+                }
+            )
+        },
+        navigationIcon = {
+            IconButton(onClick = onCancel) {
+                Icon(
+                    imageVector = Icons.Default.Close,
+                    contentDescription = "Cancel selection"
+                )
+            }
+        },
+        actions = {
+            if (selectedCount == totalCount && totalCount > 0) {
+                IconButton(onClick = onClearSelection) {
+                    Icon(
+                        imageVector = Icons.Default.Check,
+                        contentDescription = "Clear selection"
+                    )
+                }
+            } else {
+                IconButton(onClick = onSelectAll) {
+                    Icon(
+                        imageVector = Icons.Default.SelectAll,
+                        contentDescription = "Select all"
+                    )
+                }
+            }
+            IconButton(
+                onClick = onDelete,
+                enabled = selectedCount > 0
+            ) {
+                Icon(
+                    imageVector = Icons.Default.Delete,
+                    contentDescription = "Delete selected",
+                    tint = if (selectedCount > 0) {
+                        MaterialTheme.colorScheme.error
+                    } else {
+                        MaterialTheme.colorScheme.onSurface.copy(alpha = 0.38f)
+                    }
+                )
+            }
+        }
+    )
+}
