+++ app/src/main/java/com/example/stitchcounterv3/feature/navigation/NavigationRailLayout.kt
@@ -13,10 +13,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.navigation.NavHostController
-import com.example.stitchcounterv3.feature.destinations.LibraryScreenDestination
-import com.example.stitchcounterv3.feature.destinations.MainScreenDestination
-import com.example.stitchcounterv3.feature.destinations.SettingsScreenDestination
-import com.ramcosta.composedestinations.navigation.navigate
 
 @Composable
 fun NavigationRailLayout(
@@ -56,19 +52,7 @@ private fun NavigationRailComponent(
                 selected = selectedTab == tab,
                 onClick = {
                     onTabSelected(tab)
-                    val destination = when(tab) {
-                        BottomNavTab.HOME -> MainScreenDestination
-                        BottomNavTab.LIBRARY -> LibraryScreenDestination
-                        BottomNavTab.SETTINGS -> SettingsScreenDestination
-                    }
-                    navController.navigate(destination) {
-                        // Avoid building up a huge stack
-                        popUpTo(navController.graph.startDestinationId) {
-                            saveState = true
-                        }
-                        launchSingleTop = true
-                        restoreState = true
-                    }
+                    navigateToDestination(navController, getDestinationForTab(tab))
                 },
                 icon = {
                     Icon(
