+++ app/src/main/java/com/example/stitchcounterv3/feature/projectDetail/ProjectDetailScreen.kt
@@ -4,48 +4,47 @@ import android.content.Context
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.AddPhotoAlternate
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-import androidx.activity.compose.BackHandler
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
+import androidx.activity.compose.BackHandler
+import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import coil.compose.rememberAsyncImagePainter
-import coil.request.ImageRequest
 import com.example.stitchcounterv3.domain.model.ProjectType
 import com.example.stitchcounterv3.feature.navigation.RootNavGraph
 import com.example.stitchcounterv3.feature.sharedComposables.RowProgressIndicator
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
-import java.io.File
-import java.io.FileOutputStream
-import java.io.InputStream
 
 @Composable
 fun ProjectDetailContent(
@@ -221,133 +220,6 @@ fun ProjectDetailContent(
     }
 }
 
-@Composable
-private fun ProjectImageSelector(
-    imagePath: String?,
-    onImageClick: () -> Unit,
-    onRemoveImage: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Box(modifier = modifier) {
-        if (imagePath != null) {
-            ProjectImageDisplay(
-                imagePath = imagePath,
-                onImageClick = onImageClick,
-                onRemoveImage = onRemoveImage
-            )
-        } else {
-            ProjectImagePlaceholder(onImageClick = onImageClick)
-        }
-    }
-}
-
-@Composable
-private fun ProjectImageDisplay(
-    imagePath: String,
-    onImageClick: () -> Unit,
-    onRemoveImage: () -> Unit
-) {
-    val context = LocalContext.current
-    
-    Box {
-        Image(
-            painter = rememberAsyncImagePainter(
-                ImageRequest.Builder(context)
-                    .data(imagePath)
-                    .build()
-            ),
-            contentDescription = "Project image",
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(200.dp)
-                .clip(RoundedCornerShape(12.dp))
-                .clickable { onImageClick() },
-            contentScale = ContentScale.Crop
-        )
-        ProjectImageDeleteButton(
-            onRemoveImage = onRemoveImage,
-            modifier = Modifier
-                .align(Alignment.TopEnd)
-                .padding(8.dp)
-        )
-    }
-}
-
-@Composable
-private fun ProjectImageDeleteButton(
-    onRemoveImage: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    IconButton(
-        onClick = onRemoveImage,
-        modifier = modifier
-    ) {
-        Icon(
-            imageVector = Icons.Default.Delete,
-            contentDescription = "Remove image",
-            tint = MaterialTheme.colorScheme.error
-        )
-    }
-}
-
-@Composable
-private fun ProjectImagePlaceholder(
-    onImageClick: () -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onImageClick() }
-            .height(200.dp)
-            .clip(RoundedCornerShape(12.dp))
-            .background(MaterialTheme.colorScheme.primary.copy(alpha = .2f)),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Icon(
-            imageVector = Icons.Default.AddPhotoAlternate,
-            contentDescription = null,
-            modifier = Modifier.size(48.dp),
-            tint = MaterialTheme.colorScheme.tertiary
-        )
-        Text(
-            text = "Add Project Image",
-            color = MaterialTheme.colorScheme.tertiary,
-            style = MaterialTheme.typography.titleMedium,
-            textAlign = TextAlign.Center
-        )
-        Text(
-            text = "(You can add it later)",
-            color = MaterialTheme.colorScheme.tertiary,
-            style = MaterialTheme.typography.titleSmall,
-            textAlign = TextAlign.Center
-        )
-    }
-}
-
-private fun saveImageToInternalStorage(context: Context, uri: Uri, projectId: Int): String? {
-    return try {
-        val imagesDir = File(context.filesDir, "project_images")
-        if (!imagesDir.exists()) {
-            imagesDir.mkdirs()
-        }
-        
-        val fileName = "project_${projectId}_${System.currentTimeMillis()}.jpg"
-        val file = File(imagesDir, fileName)
-        
-        context.contentResolver.openInputStream(uri)?.use { inputStream: InputStream ->
-            FileOutputStream(file).use { outputStream ->
-                inputStream.copyTo(outputStream)
-            }
-        }
-        
-        file.absolutePath
-    } catch (e: Exception) {
-        android.util.Log.e("ProjectDetailScreen", "Error saving image", e)
-        null
-    }
-}
-
 @RootNavGraph
 @Destination
 @Composable
@@ -399,44 +271,4 @@ fun ProjectDetailScreen(
         onCreateProject = null,
         onNavigateBack = null
     )
-}
-
-@Composable
-private fun ProjectDetailTopBar(
-    isNewProject: Boolean,
-    onCloseClick: (() -> Unit)?,
-    onBackClick: (() -> Unit)?
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        if (onBackClick != null) {
-            IconButton(onClick = onBackClick) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = "Back"
-                )
-            }
-        } else {
-            Spacer(modifier = Modifier.size(48.dp))
-        }
-        
-        Text(
-            text = "Project Details",
-            style = MaterialTheme.typography.headlineMedium
-        )
-        
-        if (onCloseClick != null) {
-            IconButton(onClick = onCloseClick) {
-                Icon(
-                    imageVector = Icons.Default.Close,
-                    contentDescription = "Close"
-                )
-            }
-        } else {
-            Spacer(modifier = Modifier.size(48.dp))
-        }
-    }
 }
