+++ app/src/main/java/com/example/stitchcounterv3/feature/sharedComposables/CounterTopBar.kt
@@ -1,0 +1,41 @@
+package com.example.stitchcounterv3.feature.sharedComposables
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun CounterTopBar(
+    title: String,
+    topBarContent: (@Composable () -> Unit)? = null,
+    modifier: Modifier = Modifier
+) {
+    if (title.isNotEmpty() || topBarContent != null) {
+        Row(
+            modifier = modifier
+                .fillMaxWidth()
+                .padding(bottom = 8.dp),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            if (title.isNotEmpty()) {
+                Text(
+                    text = title,
+                    style = MaterialTheme.typography.titleMedium,
+                    modifier = Modifier.weight(1f)
+                )
+            } else {
+                Spacer(modifier = Modifier.weight(1f))
+            }
+            topBarContent?.invoke()
+        }
+    }
+}
