+++ app/src/main/java/com/example/stitchcounterv3/feature/sharedComposables/ProjectDetailsFAB.kt
@@ -1,0 +1,32 @@
+package com.example.stitchcounterv3.feature.sharedComposables
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ProjectDetailsFAB(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    FloatingActionButton(
+        onClick = onClick,
+        modifier = modifier
+            .padding(start = 16.dp)
+            .size(40.dp),
+        containerColor = MaterialTheme.colorScheme.primaryContainer
+    ) {
+        Icon(
+            imageVector = Icons.Default.Info,
+            contentDescription = "Project details",
+            modifier = Modifier.size(20.dp)
+        )
+    }
+}
